+++ lib/app/service/connection/mapping/connection_network_mapper.dart
@@ -1,7 +1,7 @@
 import 'package:app/app/service/connection/data/connection_network/connection_network.dart';
 import 'package:app/app/service/connection/mapping/custom_network_options_mapper.dart';
 import 'package:app/app/service/connection/mapping/networks_mapper.dart';
-import 'package:app/app/service/connection/mapping/tansports_mapper.dart';
+import 'package:app/app/service/connection/mapping/transports_mapper.dart';
 import 'package:app/utils/json/json_utils.dart';
 
 ConnectionNetwork mapToConnectionNetworkFromJson(Map<String, dynamic> json) {
